requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
aiohttp>=3.9.3
beautifulsoup4>=4.12.3
pyyaml>=6.0.1
unidecode>=1.3.8
python-dotenv>=1.0.1
tabula-py>=2.9.0
pytesseract>=0.3.10
Pillow>=10.1.0
docx2txt>=0.8
pdfplumber>=0.10.3
openai>=1.12.0
openpyxl>=3.1.5
plotly>=5.18.0

# DocLing y sus dependencias
docling>=2.24.0
docling-core[chunking]>=2.20.0
docling-ibm-models>=3.4.0
docling-parse>=3.4.0
torch>=2.6.0
torchvision>=0.21.0
transformers>=4.49.0
scipy>=1.15.2
lxml>=5.3.1
rtree>=1.3.0
typer>=0.12.5
pydantic>=2.10.6
pydantic-settings>=2.8.0
